// cocktails.web/src/api/cocktailsApi/models/cocktailIngredientFiltersRs.ts
export interface CocktailIngredientFiltersRs {
    /**
     * The cocktail ingredient filters for searching against recommended glassware
     * @type {Array<IngredientFilterModel>}
     * @memberof CocktailIngredientFiltersRs
     */
    glassware: Array<IngredientFilterModel>;
    /**
     * The cocktail ingredient filters for searching against spirits
     * @type {Array<any>}
     * @memberof CocktailIngredientFiltersRs
     */
    spirits: Array<any>;
    /**
     * The cocktail ingredient filters for searching against liqueurs
     * @type {Array<any>}
     * @memberof CocktailIngredientFiltersRs
     */
    liqueurs: Array<any>;
    /**
     * The cocktail ingredient filters for searching against friuts
     * @type {Array<any>}
     * @memberof CocktailIngredientFiltersRs
     */
    fruits: Array<any>;
    /**
     * The cocktail ingredient filters for searching against vegetables
     * @type {Array<any>}
     * @memberof CocktailIngredientFiltersRs
     */
    vegetables: Array<any>;
    /**
     * The cocktail ingredient filters for searching against herbs and flowers
     * @type {Array<any>}
     * @memberof CocktailIngredientFiltersRs
     */
    herbsAndFlowers: Array<any>;
    /**
     * The cocktail ingredient filters for searching against syrups and sauces
     * @type {Array<any>}
     * @memberof CocktailIngredientFiltersRs
     */
    syrupsAndSauces: Array<any>;
    /**
     * The cocktail ingredient filters for searching against bitters
     * @type {Array<any>}
     * @memberof CocktailIngredientFiltersRs
     */
    bitters: Array<any>;
    /**
     * The cocktail ingredient filters for searching against proteins
     * @type {Array<any>}
     * @memberof CocktailIngredientFiltersRs
     */
    proteins: Array<any>;
    /**
     * The cocktail ingredient filters for searching against juices
     * @type {Array<any>}
     * @memberof CocktailIngredientFiltersRs
     */
    juices: Array<any>;
    /**
     * The cocktail ingredient filters for searching against dilutions
     * @type {Array<any>}
     * @memberof CocktailIngredientFiltersRs
     */
    dilutions: Array<any>;
    /**
     * The cocktail ingredient filters for searching against beers, wines and champagnes
     * @type {Array<any>}
     * @memberof CocktailIngredientFiltersRs
     */
    beerWineChampagne: Array<any>;
    /**
     * The cocktail ingredient filters for searching against eras when cocktails were established
     * @type {Array<any>}
     * @memberof CocktailIngredientFiltersRs
     */
    eras: Array<any>;
}